backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
httpx==0.26.0

# LLM Integration
openai==1.10.0
anthropic==0.18.0

# Payment (YooKassa)
yookassa==3.1.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# CORS
starlette==0.35.1
